green color theme variables
